rename package 'k8s'->'fs'
