+++ src/main/java/org/zalando/stups/tokens/fs/AccessTokenDto.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.zalando.stups.tokens.k8s;
+package org.zalando.stups.tokens.fs;
 
 import org.zalando.stups.tokens.AccessToken;
